# saas-forge-contracts/compatibility-baselines/v1/v0.1.0/openapi/v1.yaml
openapi: 3.1.1
info:
  title: saas-forge Public API
  version: 1.0.0
  description: |
    The versioned public REST contract for saas-forge. This document contains only
    delivery-stage-two and delivery-stage-three resources. Permission, Feature,
    and Quota runtime operations are intentionally absent until their domain
    contracts are reviewed.
servers:
  - url: https://api.example.invalid
    description: Deployment-specific API Gateway origin.
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      operationId: refreshAccessToken
      x-saas.forge-service: iam-service
      summary: Rotate the browser Refresh Token and issue the current Access Token
      description: The Refresh Token is supplied only by the HttpOnly refresh cookie.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyRequest'
      responses:
        '200':
          description: The Refresh Token was rotated.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/RefreshCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResult'
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
...
  /api/v1/auth/tenant-switches:
    post:
      tags: [Authentication]
      operationId: switchTenantContext
      x-saas.forge-service: iam-service
      summary: Select or switch the session Membership
      description: |
        Accepts only the target Membership. The current Identity is authenticated by
        either its Access Token or Refresh Token cookie; IAM verifies ownership and
        Tenant accessibility before changing the session. Call refresh after 204.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantSwitchRequest'
      responses:
        '204': { $ref: './common.yaml#/components/responses/NoContent' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
  /oauth2/token:
    post:
      tags: [Authentication]
      operationId: issueClientCredentialsToken
      x-saas.forge-service: iam-service
      summary: Issue a Client Credentials Access Token
      description: |
        Supports only grant_type=client_credentials. The OAuth Client authenticates
        with HTTP Basic; requested scope must be a subset of its allowedScopes.
      security:
        - OAuthClientBasic: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ClientCredentialsTokenRequest'
      responses:
        '200':
          description: A service Access Token was issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientCredentialsTokenResponse'
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
...
components:
  securitySchemes:
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Platform Admin user Access Token. OAuth Client tokens cannot access platform management operations.
    OAuthClientBasic:
      type: http
      scheme: basic
      description: OAuth Client client_id and client_secret.
  headers:
    RefreshCookie:
      description: A rotated __Host-sf_refresh HttpOnly, Secure, SameSite=Strict cookie scoped to Path=/.
      required: true
      schema: { type: string }
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { $ref: './common.yaml#/components/schemas/UuidV7' }
    CsrfHeader:
      name: X-SF-CSRF
      in: header
      required: true
      schema:
        type: string
        const: '1'
...
  schemas:
...
    EmptyRequest:
      type: object
      additionalProperties: false
...
    AccessTokenResult:
      type: object
      required: [contextState, accessToken, tokenType, expiresIn]
      properties:
        contextState: { const: ACCESS_TOKEN_ISSUED }
        accessToken: { type: string, minLength: 1 }
        tokenType: { type: string, const: Bearer }
        expiresIn: { type: integer, minimum: 1 }
      additionalProperties: false
    MembershipCandidate:
      type: object
      required: [membershipId, tenantId, tenantDisplayName]
      properties:
        membershipId: { $ref: './common.yaml#/components/schemas/UuidV7' }
        tenantId: { $ref: './common.yaml#/components/schemas/UuidV7' }
        tenantDisplayName: { type: string, minLength: 1, maxLength: 200 }
      additionalProperties: false
    ContextSelectionRequiredResult:
      type: object
      required: [contextState, memberships]
      properties:
        contextState: { const: CONTEXT_SELECTION_REQUIRED }
        memberships:
          type: array
          minItems: 2
          items: { $ref: '#/components/schemas/MembershipCandidate' }
      additionalProperties: false
    AuthenticationResult:
      oneOf:
        - $ref: '#/components/schemas/AccessTokenResult'
        - $ref: '#/components/schemas/ContextSelectionRequiredResult'
    TenantSwitchRequest:
      type: object
      required: [membershipId]
      properties:
        membershipId: { $ref: './common.yaml#/components/schemas/UuidV7' }
      additionalProperties: false
    ClientCredentialsTokenRequest:
      type: object
      required: [grant_type]
      properties:
        grant_type: { type: string, const: client_credentials }
        scope:
          type: string
          description: Space-delimited requested scopes, each of which must have been granted to the OAuth Client.
      additionalProperties: false
    ClientCredentialsTokenResponse:
      type: object
      required: [access_token, token_type, expires_in, scope]
      properties:
        access_token: { type: string, minLength: 1 }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, minimum: 1 }
        scope:
          type: string
          description: Space-delimited scopes granted to this token.
      additionalProperties: false
...